requirements/testing.txt
-r base.txt

celery==3.1.17
django-jenkins
coverage
mock==1.0.1

redgreenunittest==0.1.1

django-behave
selenium
